+++ components/forms/farmer-journal/Note.tsx
